#60 - refactoring: moved jobs to new package
