+++ src/main/java/de/bayern/gdi/gui/controller/ButtonBarController.java
@@ -21,7 +21,7 @@ import de.bayern.gdi.gui.ProgressDialog;
 import de.bayern.gdi.model.DownloadStep;
 import de.bayern.gdi.processor.ConverterException;
 import de.bayern.gdi.processor.DownloadStepConverter;
-import de.bayern.gdi.processor.JobList;
+import de.bayern.gdi.processor.job.JobList;
 import de.bayern.gdi.processor.Processor;
 import de.bayern.gdi.utils.Config;
 import de.bayern.gdi.utils.I18n;
