+++ src/main/java/de/bayern/gdi/processor/ProcessingStepConverter.java
@@ -28,10 +28,14 @@ import de.bayern.gdi.model.DownloadStep;
 import de.bayern.gdi.model.ProcessingConfiguration;
 import de.bayern.gdi.model.ProcessingStep;
 import de.bayern.gdi.model.ProcessingStepConfiguration;
-import de.bayern.gdi.processor.ExternalProcessJob.Arg;
-import de.bayern.gdi.processor.ExternalProcessJob.DeltaGlob;
-import de.bayern.gdi.processor.ExternalProcessJob.GlobalGlob;
-import de.bayern.gdi.processor.ExternalProcessJob.UniqueArg;
+import de.bayern.gdi.processor.job.BroadcastJob;
+import de.bayern.gdi.processor.job.ExternalProcessJob;
+import de.bayern.gdi.processor.job.ExternalProcessJob.Arg;
+import de.bayern.gdi.processor.job.ExternalProcessJob.DeltaGlob;
+import de.bayern.gdi.processor.job.ExternalProcessJob.GlobalGlob;
+import de.bayern.gdi.processor.job.ExternalProcessJob.UniqueArg;
+import de.bayern.gdi.processor.job.GmlEmptyCheckJob;
+import de.bayern.gdi.processor.job.Job;
 import de.bayern.gdi.utils.Config;
 import de.bayern.gdi.utils.FileTracker;
 import de.bayern.gdi.utils.I18n;
