+++ src/main/java/de/bayern/gdi/processor/job/AtomDownloadJob.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.bayern.gdi.processor;
+package de.bayern.gdi.processor.job;
 
 import java.io.File;
 import java.io.IOException;
@@ -26,6 +26,7 @@ import java.util.HashMap;
 import javax.xml.namespace.NamespaceContext;
 import javax.xml.xpath.XPathConstants;
 
+import de.bayern.gdi.processor.JobExecutionException;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.w3c.dom.Document;
