+++ src/main/java/de/bayern/gdi/processor/job/FilesDownloadJob.java
@@ -15,13 +15,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.bayern.gdi.processor;
+package de.bayern.gdi.processor.job;
 
 import java.io.File;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 
+import de.bayern.gdi.processor.JobExecutionException;
 import org.apache.http.HttpEntity;
 
 /** Download a list of URLs to given files. */
