+++ src/main/java/de/bayern/gdi/processor/job/LogMetaJob.java
@@ -15,13 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.bayern.gdi.processor;
+package de.bayern.gdi.processor.job;
 
 import java.util.List;
 
 import de.bayern.gdi.model.Parameter;
 import de.bayern.gdi.model.DownloadStep;
 
+import de.bayern.gdi.processor.JobExecutionException;
+import de.bayern.gdi.processor.Processor;
 import de.bayern.gdi.utils.I18n;
 import de.bayern.gdi.utils.Log;
