+++ src/main/java/de/bayern/gdi/processor/job/UnzipJob.java
@@ -15,8 +15,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.bayern.gdi.processor;
+package de.bayern.gdi.processor.job;
 
+import de.bayern.gdi.processor.JobExecutionException;
+import de.bayern.gdi.processor.Processor;
 import de.bayern.gdi.utils.FileTracker;
 import de.bayern.gdi.utils.I18n;
 import de.bayern.gdi.utils.Log;
@@ -44,7 +46,7 @@ public class UnzipJob implements Job {
      * @param fileTracker never <code>null</code>
      * @param logger      never <code>null</code>
      */
-    UnzipJob(FileTracker fileTracker, Log logger) {
+    public UnzipJob(FileTracker fileTracker, Log logger) {
         this.fileTracker = fileTracker;
         this.logger = logger;
     }
