+++ src/test/java/de/bayern/gdi/processor/job/GmlEmptyCheckJobTest.java
@@ -15,8 +15,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.bayern.gdi.processor;
+package de.bayern.gdi.processor.job;
 
+import de.bayern.gdi.processor.JobExecutionException;
+import de.bayern.gdi.processor.Processor;
 import de.bayern.gdi.utils.FileTracker;
 import org.apache.commons.io.IOUtils;
 import org.junit.Rule;
