+++ src/test/java/de/bayern/gdi/processor/job/UnzipJobTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.bayern.gdi.processor;
+package de.bayern.gdi.processor.job;
 
 import de.bayern.gdi.utils.FileTracker;
 import de.bayern.gdi.utils.Log;
