+++ src/test/resources/de/bayern/gdi/processor/job/emptyFeatureCollection.gml
