+++ src/test/resources/de/bayern/gdi/processor/job/zippedFile.zip
